refactor: ViewMode/PctDirection/TableOpts/ChartOptsをviewTypes.tsに集約
Toolbar.tsxに定義されていたViewMode/PctDirectionと、ResultCard.tsxに
定義されていたTableOpts/ChartOptsをviewTypes.tsに一元化。
ViewOptsPropsでもChartOptsを再利用しprops数を削減。
